Added 5 new projects for the solution structuring the architecture of the solution:
DataAccess - accessing the db containing the given DataAccessLayer class.

DataModel - mapping the data into objects containing the given base class ToDo.

BusinessLogic – containing the business logic for the application that is used by the service.

Service – will expose the functionality of the application to the consumer.

Client – used to simulate a system or user consuming the service.